+++ app/src/main/java/com/jin/jjinweather/layer/domain/model/weather/HourlyWeather.kt
@@ -3,5 +3,5 @@ package com.jin.jjinweather.layer.domain.model.weather
 data class HourlyWeather(
     val forecastTime: Long,
     val iconResId: Int,
-    val temperature: Double
+    val temperature: Int
 )
